Addressing comments from fourth round of review
