+++ storage/src/main/java/org/apache/kafka/storage/internals/log/LogConfig.java
@@ -547,7 +547,7 @@ public class LogConfig extends AbstractConfig {
      * The default values should be extracted from the KafkaConfig.
      * @param props The properties to be validated
      */
-    public static void validateTopicLogConfigValues(Map<?, ?> props,
+    private static void validateTopicLogConfigValues(Map<?, ?> props,
                                                     boolean isRemoteLogStorageSystemEnabled) {
         validateValues(props);
         boolean isRemoteLogStorageEnabled = (Boolean) props.get(TopicConfig.REMOTE_LOG_STORAGE_ENABLE_CONFIG);
@@ -563,7 +563,7 @@ public class LogConfig extends AbstractConfig {
         boolean isRemoteLogStorageEnabled = (Boolean) props.get(TopicConfig.REMOTE_LOG_STORAGE_ENABLE_CONFIG);
         if (isRemoteLogStorageEnabled && !isRemoteLogStorageSystemEnabled) {
             if (isReceivingConfigFromStore) {
-                throw new ConfigException("You have to disable tiering on all topics with the property remote.storage.enable=true before disabling tiered storage cluster-wide");
+                throw new ConfigException("You have to delete all topics with the property remote.storage.enable=true before disabling tiered storage cluster-wide");
             } else {
                 throw new ConfigException("Tiered Storage functionality is disabled in the broker. " +
                         "Topic cannot be configured with remote log storage.");
